Extract Countable & Weightable interfaces from ReactionCounter & ReactionTotal models
