+++ src/Reactant/ReactionTotal/Models/ReactionTotal.php
@@ -14,20 +14,12 @@ declare(strict_types=1);
 namespace I\Love\Reactant\ReactionTotal\Models;
 
 use I\Love\Reactant\Models\Reactant;
+use I\Love\Support\Object\Countable;
+use I\Love\Support\Object\Weightable;
 
-interface ReactionTotal
+interface ReactionTotal extends
+    Countable,
+    Weightable
 {
     public function getReactant(): Reactant;
-
-    public function getCount(): int;
-
-    public function getWeight(): float;
-
-    public function incrementCount(int $amount): void;
-
-    public function decrementCount(int $amount): void;
-
-    public function incrementWeight(float $amount): void;
-
-    public function decrementWeight(float $amount): void;
 }
